requirements.txt
absl-py==2.3.1
matplotlib==3.10.7
numpy==2.3.2
pandas==2.3.2
PyQt6==6.9.1
PyQt6-Qt6==6.9.2
PyQt6_sip==13.10.2
scikit-learn==1.7.1
scipy==1.16.1
silence_tensorflow==1.2.3
tensorflow==2.20.0
